backend/old/resume_agent/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
groq==0.31.0
pydantic==2.11.7
python-dotenv==1.1.1
azure-ai-formrecognizer==3.3.3